backend/requirements.txt
Django==3.2.11
djangorestframework==3.12.4
psycopg2==2.9.1
django-dbconn-retry

